Move all the types to /types directory
