+++ src/components/VAllResultsGrid/VAudioCell.vue
@@ -5,7 +5,7 @@
 <script lang="ts">
 import { defineComponent, PropType } from '@nuxtjs/composition-api'
 
-import type { AudioDetail } from '~/models/media'
+import type { AudioDetail } from '~/types/media'
 
 import VAudioTrack from '~/components/VAudioTrack/VAudioTrack.vue'
