+++ src/components/VAudioDetails/VAudioDetails.vue
@@ -88,7 +88,7 @@
 <script lang="ts">
 import { computed, defineComponent, PropType } from '@nuxtjs/composition-api'
 
-import type { AudioDetail } from '~/models/media'
+import type { AudioDetail } from '~/types/media'
 
 import VAudioThumbnail from '~/components/VAudioThumbnail/VAudioThumbnail.vue'
 import VContentReportPopover from '~/components/VContentReport/VContentReportPopover.vue'
