+++ src/components/VAudioTrack/VGlobalAudioTrack.vue
@@ -41,7 +41,7 @@ import { useI18n } from '~/composables/use-i18n'
 import { useActiveMediaStore } from '~/stores/active-media'
 import { useMediaStore } from '~/stores/media'
 
-import type { AudioDetail } from '~/models/media'
+import type { AudioDetail } from '~/types/media'
 import type { AudioStatus } from '~/constants/audio'
 
 import VPlayPause from '~/components/VAudioTrack/VPlayPause.vue'
