+++ src/components/VAudioTrack/layouts/VBoxLayout.vue
@@ -40,7 +40,7 @@
 <script lang="ts">
 import { computed, defineComponent, PropType } from '@nuxtjs/composition-api'
 
-import type { AudioDetail } from '~/models/media'
+import type { AudioDetail } from '~/types/media'
 import type { AudioSize } from '~/constants/audio'
 import { useI18n } from '~/composables/use-i18n'
