+++ src/components/VImageDetails/VRelatedImages.vue
@@ -15,8 +15,8 @@
 <script lang="ts">
 import { defineComponent, PropType } from '@nuxtjs/composition-api'
 
-import type { ImageDetail } from '~/models/media'
-import type { FetchState } from '~/models/fetch-state'
+import type { ImageDetail } from '~/types/media'
+import type { FetchState } from '~/types/fetch-state'
 
 import VImageGrid from '~/components/VImageGrid/VImageGrid.vue'
 import VLoadingIcon from '~/components/LoadingIcon.vue'
