+++ src/pages/image/_id.vue
@@ -93,7 +93,7 @@ import {
 } from '@nuxtjs/composition-api'
 
 import { IMAGE } from '~/constants/media'
-import type { ImageDetail } from '~/models/media'
+import type { ImageDetail } from '~/types/media'
 import { useSingleResultStore } from '~/stores/media/single-result'
 import { useRelatedMediaStore } from '~/stores/media/related-media'
 import { createDetailPageMeta } from '~/utils/og'
