+++ src/pages/search/search-page.types.ts
@@ -1,6 +1,6 @@
 import type { SupportedMediaType } from '~/constants/media'
-import type { Media } from '~/models/media'
-import type { FetchState } from '~/models/fetch-state'
+import type { Media } from '~/types/media'
+import type { FetchState } from '~/types/fetch-state'
 
 import type { ExtractPropTypes, PropType } from '@nuxtjs/composition-api'
