+++ src/stores/media/index.ts
@@ -5,8 +5,8 @@ import axios from 'axios'
 import { warn } from '~/utils/console'
 import { hash, rand as prng } from '~/utils/prng'
 import prepareSearchQueryParams from '~/utils/prepare-search-query-params'
-import type { DetailFromMediaType, Media } from '~/models/media'
-import type { FetchState } from '~/models/fetch-state'
+import type { DetailFromMediaType, Media } from '~/types/media'
+import type { FetchState } from '~/types/fetch-state'
 import {
   ALL_MEDIA,
   AUDIO,
