+++ src/stores/media/services.ts
@@ -1,6 +1,6 @@
 import { AUDIO, IMAGE } from '~/constants/media'
 import MediaService from '~/data/media-service'
-import type { AudioDetail, ImageDetail } from '~/models/media'
+import type { AudioDetail, ImageDetail } from '~/types/media'
 import { createApiService } from '~/data/api-service'
 
 export const initServices = {
