+++ src/types/fetch-state.ts
